\name{pathway}
\alias{pathway}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Example dataset for fasting and covid19 datasets}
\description{
This data includes a metabolite set list and metabolite name list for fasting, and a metabolite set list for covid19 dataset within the "loadings" package
}
\usage{
data(pathway)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
The list object pathway contains the following elements:

  fasting : 	metabolite set list for fasting mouse dataset

  data$fasting : metabolite name list for fasting mouse dataset

  covid19$proteomics :   protein set list for covid19 dataset.
}
\references{
Yamamoto H., Fujimori T., Sato H., Ishikawa G., Kami K., Ohashi Y. (2014). "Statistical hypothesis testing of factor loading in principal component analysis and its application to metabolite set enrichment analysis". BMC Bioinformatics, (2014) 15(1):51.

B. Shen, et al, Proteomic and Metabolomic Characterization of COVID-19 Patient Sera, Cell. 182 (2020) 59-72.e15.
}
\examples{
data(pathway)
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ datasets }
